Phase 4: Remove dead code, fix naming, consolidate notifications
- Delete Date+Format.swift (146 lines, zero call sites)
- Delete Reach.swift (69 lines, replaced with inline SCNetworkReachability)
- Rename Panelr → CustomPanel (fix typo in class name)
- Consolidate duplicate theme notification names into single
  .themeDidChangeNotification (was defined twice as themeDidChange
  and themeDidChangeNotification, both resolving to same string)
- Update all XIB/storyboard references for CustomPanel rename
- Update all notification subscribers to use consolidated name
