+++ Clocker/Overall App/NetworkManager.swift
@@ -1,6 +1,7 @@
 // Copyright © 2015 Abhishek Banthia
 
 import Cocoa
+import SystemConfiguration
 
 class NetworkManager: NSObject {
     static let parsingError: NSError = {
@@ -161,14 +162,23 @@ extension NetworkManager {
             return false
         }
 
-        let status = Reach().connectionStatus()
-        switch status {
-        case .online(.wwan):
-            return true
-        case .online(.wiFi):
-            return true
-        default:
+        var zeroAddress = sockaddr_in()
+        zeroAddress.sin_len = UInt8(MemoryLayout.size(ofValue: zeroAddress))
+        zeroAddress.sin_family = sa_family_t(AF_INET)
+
+        guard let reachability = withUnsafePointer(to: &zeroAddress, {
+            $0.withMemoryRebound(to: sockaddr.self, capacity: 1) {
+                SCNetworkReachabilityCreateWithAddress(nil, $0)
+            }
+        }) else {
             return false
         }
+
+        var flags: SCNetworkReachabilityFlags = []
+        if !SCNetworkReachabilityGetFlags(reachability, &flags) {
+            return false
+        }
+
+        return flags.contains(.reachable) && !flags.contains(.connectionRequired)
     }
 }
